+++ backend/app/api/routes/pentester.py
@@ -1,0 +1,43 @@
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, status
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.api.middleware.auth import require_pentester
+from app.models.user import User
+from app.schemas.engagement import PentesterConversationListResponse
+from app.schemas.finding import FindingCreate, FindingListItem
+from app.services.engagement_service import EngagementService
+from app.utils.db import get_db
+
+router = APIRouter(prefix="/pentester", tags=["Pentester"])
+
+
+@router.get("/messages", response_model=PentesterConversationListResponse)
+async def get_pentester_messages(
+    db: AsyncSession = Depends(get_db),
+    user: User = Depends(require_pentester),
+) -> PentesterConversationListResponse:
+    return await EngagementService.get_message_conversations(
+        db,
+        user_id=user.id,
+    )
+
+
+@router.post(
+    "/{engagement_id}/findings", 
+    response_model=FindingListItem, 
+    status_code=status.HTTP_201_CREATED,
+)
+async def create_engagement_finding(
+    engagement_id: UUID,
+    request: FindingCreate,
+    db: AsyncSession = Depends(get_db),
+    user: User = Depends(require_pentester),
+) -> FindingListItem:
+    return await EngagementService.create_manual_finding(
+        db,
+        engagement_id=engagement_id,
+        user_id=user.id,
+        request=request,
+    )
